Module tch::utils

Enums

QEngine

Quantization engines

Functions

get_num_interop_threads

Get the number of threads used by torch for inter-op parallelism.

get_num_threads

Get the number of threads used by torch in parallel regions.

has_cuda

has_cudart

has_cusolver

has_hip

has_ipu

has_lapack

has_lazy

has_magma

has_mkl

has_mkldnn

has_mps

has_openmp

has_ort

has_vulkan

Check whether the vulkan backend is available. None that this backend is not included by default as of PyTorch 2.0.0. https://pytorch.org/tutorials/prototype/vulkan_workflow.html#building-pytorch-with-vulkan-backend

has_xla

manual_seed

Sets the random seed used by torch.

set_num_interop_threads

Set the number of threads used by torch for inter-op parallelism.

set_num_threads

Set the number of threads used by torch in parallel regions.

version_cudart

version_cudnn