Module tch::vision

The vision module groups functions and models related to computer vision.

Modules

alexnet

AlexNet. https://arxiv.org/abs/1404.5997

cifar

The CIFAR-10 dataset.

convmixer

ConvMixer implementation.

dataset

A simple dataset structure shared by various computer vision datasets.

densenet

DenseNet implementation.

dinov2

DINOv2: Learning Robust Visual Features without Supervision https://github.com/facebookresearch/dinov2 The weights can be extracted from pre-trained Python models using python src/vision/export_dinov2.py.

efficientnet

EfficientNet implementation.

image

Utility functions to manipulate images.

imagenet

Helper functions for ImageNet like datasets.

inception

InceptionV3.

mnist

The MNIST hand-written digit dataset.

mobilenet

MobileNet V2 implementation. https://ai.googleblog.com/2018/04/mobilenetv2-next-generation-of-on.htmla

resnet

ResNet implementation.

squeezenet

SqueezeNet implementation.

vgg

VGG models.